+++ beaker-ts/.gitignore
@@ -55,3 +55,5 @@ build
 # MISC
 create.sql
 .eslintcache
+
+beaker-kernel-[0-9\.]*.tgz
