+++ beaker-vue/.gitignore
@@ -29,6 +29,8 @@ coverage
 
 *.tsbuildinfo
 
+beaker-vue-[0-9\.]*.tgz
+
 test-results/
 playwright-report/
 html/
